requirements.txt
PyQt5==5.15.10
PyMuPDF==1.24.1
PyPDF2==3.0.1
pdf2image==1.17.0
python-docx==0.8.11
docx2pdf==0.1.8
Pillow==10.3.0
reportlab==4.1.0
pikepdf==8.13.0
opencv-python==4.9.0.80
pytesseract==0.3.10
tqdm==4.66.4
pandas==2.2.2
openpyxl==3.1.2
python-pptx==0.6.23
